Minor change to trigger better prioritization
Testing revealed that the system can get fixated on an aircraft, and would only lose track once the aircraft had gone out of range of the SDR.  This is not ideal, we want a track to happen for the most optimal aircraft for photos, with some stickiness to prevent snapping around on multiple tracks.  This code change will cause the system to re-evaluate for an optimal target every OBSERVATION_CLEAN_INTERVAL, currently every 30 seconds by default.


Former-commit-id: b234d48db438f0f3316686e8c97d973479a14c57